fix false-positive accumulation resets at grazing angles and stripes on far-away geometry

+++ shaders/lib/pipelineSettings.glsl
@@ -1,7 +1,7 @@
 /*
 const int colortex0Format = R11F_G11F_B10F;	//main color
 const int colortex1Format = RGBA8;			//smoothnessD & materialMask & skyLightFactor
-const int colortex2Format = RGBA16;			//taa, inverted depth
+const int colortex2Format = RGBA16F;			//taa, inverted depth
 const int colortex3Format = RGBA8;		    //(cloud/water map on deferred) | translucentMult & bloom & final color // can replace colortex8
 const int colortex4Format = RGBA8;			//volumetric cloud linear depth & volumetric light factor & normalM in composite
 const int colortex5Format = RGBA8_SNORM;	//normalM & scene image for water reflections
